created modal pop up for notifcation bell + moved travelboard notification to bell #195 #197
